Finished Renaming and Breaking Out Classes
* Renamed files so Request Method was in front of `Handler`
* Broke out the FileReaderHandler
